🕸️ Weaver: Refactor KnowledgePage using Custom Hook Pattern
