+++ src/features/knowledge/KnowledgePage.tsx
@@ -1,12 +1,4 @@
-import {
-  memo,
-  useCallback,
-  useDeferredValue,
-  useEffect,
-  useMemo,
-  useState,
-  useTransition,
-} from 'react';
+import { memo, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
   Database,
@@ -17,7 +9,6 @@ import {
   Search,
   Trash2,
 } from 'lucide-react';
-import { toast } from 'sonner';
 import {
   Badge,
   Button,
@@ -54,19 +45,13 @@ import {
 } from '@/components/ui/alert-dialog';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Skeleton } from '@/components/ui/skeleton';
-import { getLogger } from '@/lib/logger';
 import {
-  deleteGlobalKnowledge,
-  getGlobalKnowledgeDetail,
-  listGlobalKnowledge,
   type KnowledgeChunkDetail,
   type KnowledgeChunkListItem,
   type KnowledgeGraphEntity,
-  type KnowledgeListCursor,
 } from '@/lib/backend/knowledge';
+import { useGlobalKnowledge } from './hooks/useGlobalKnowledge';
 
-const logger = getLogger('KnowledgePage');
-const KNOWLEDGE_PAGE_SIZE = 60;
 const knowledgeDateFormatter = new Intl.DateTimeFormat(undefined, {
   dateStyle: 'medium',
   timeStyle: 'short',
@@ -576,229 +561,32 @@ const KnowledgeDetailDialog = memo(function KnowledgeDetailDialog({
 
 export default function KnowledgePage() {
   const { t } = useTranslation('common');
-  const [query, setQuery] = useState('');
-  const [debouncedQuery, setDebouncedQuery] = useState('');
-  const [assistantFilter, setAssistantFilter] = useState('all');
-  const [items, setItems] = useState<KnowledgeChunkListItem[]>([]);
-  const [assistants, setAssistants] = useState<string[]>([]);
-  const [selectedId, setSelectedId] = useState<number | null>(null);
-  const [detail, setDetail] = useState<KnowledgeChunkDetail | null>(null);
-  const [isListLoading, setIsListLoading] = useState(true);
-  const [isLoadingMore, setIsLoadingMore] = useState(false);
-  const [isDetailLoading, setIsDetailLoading] = useState(false);
-  const [isDeleting, setIsDeleting] = useState(false);
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
-  const [listRefreshToken, setListRefreshToken] = useState(0);
-  const [nextCursor, setNextCursor] = useState<KnowledgeListCursor | null>(
-    null,
-  );
-  const deferredQuery = useDeferredValue(query);
-  const [, startSelectionTransition] = useTransition();
-
-  useEffect(() => {
-    const timeout = window.setTimeout(() => {
-      setDebouncedQuery(deferredQuery.trim());
-    }, 250);
-
-    return () => window.clearTimeout(timeout);
-  }, [deferredQuery]);
-
-  useEffect(() => {
-    let cancelled = false;
-
-    const load = async () => {
-      setIsListLoading(true);
-      try {
-        const response = await listGlobalKnowledge({
-          query: debouncedQuery || undefined,
-          assistantId: assistantFilter === 'all' ? undefined : assistantFilter,
-          limit: KNOWLEDGE_PAGE_SIZE,
-        });
-
-        if (cancelled) {
-          return;
-        }
-
-        setItems(response.items);
-        setAssistants(response.assistants);
-        setNextCursor(response.nextCursor ?? null);
-        setSelectedId((current) => {
-          if (current && response.items.some((item) => item.id === current)) {
-            return current;
-          }
-          return null;
-        });
-      } catch (error) {
-        logger.error('Failed to load global knowledge list', error);
-        if (!cancelled) {
-          toast.error(
-            t(
-              'knowledge.loadListFailed',
-              'Failed to load global knowledge entries.',
-            ),
-          );
-        }
-      } finally {
-        if (!cancelled) {
-          setIsListLoading(false);
-        }
-      }
-    };
-
-    void load();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [assistantFilter, debouncedQuery, listRefreshToken, t]);
-
-  useEffect(() => {
-    if (selectedId === null) {
-      setDetail(null);
-      return;
-    }
-
-    let cancelled = false;
-
-    const loadDetail = async () => {
-      setIsDetailLoading(true);
-      try {
-        const response = await getGlobalKnowledgeDetail(selectedId);
-        if (!cancelled) {
-          setDetail(response);
-        }
-      } catch (error) {
-        logger.error('Failed to load knowledge detail', { selectedId, error });
-        if (!cancelled) {
-          toast.error(
-            t(
-              'knowledge.loadDetailFailed',
-              'Failed to load knowledge details.',
-            ),
-          );
-          setDetail(null);
-        }
-      } finally {
-        if (!cancelled) {
-          setIsDetailLoading(false);
-        }
-      }
-    };
-
-    void loadDetail();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [selectedId, t]);
-
-  const selectedItem = useMemo(
-    () => items.find((item) => item.id === selectedId) ?? null,
-    [items, selectedId],
-  );
-  const isDetailOpen = selectedItem !== null;
-  const hasMoreItems = nextCursor !== null;
-  const entityNameById = useMemo(
-    () =>
-      new Map(detail?.entities.map((entity) => [entity.id, entity.name]) ?? []),
-    [detail?.entities],
-  );
-
-  const handleRefresh = useCallback(() => {
-    setListRefreshToken((current) => current + 1);
-  }, []);
-
-  const handleCloseDetail = useCallback(() => {
-    setIsDeleteDialogOpen(false);
-    setSelectedId(null);
-  }, []);
-
-  const handleSelectItem = useCallback((id: number) => {
-    startSelectionTransition(() => {
-      setSelectedId(id);
-    });
-  }, []);
-
-  const handleLoadMore = useCallback(async () => {
-    if (isListLoading || isLoadingMore || nextCursor === null) {
-      return;
-    }
-
-    setIsLoadingMore(true);
-    try {
-      const response = await listGlobalKnowledge({
-        query: debouncedQuery || undefined,
-        assistantId: assistantFilter === 'all' ? undefined : assistantFilter,
-        cursor: nextCursor,
-        limit: KNOWLEDGE_PAGE_SIZE,
-      });
-
-      setItems((current) => {
-        const seenIds = new Set(current.map((item) => item.id));
-        const appendedItems = response.items.filter(
-          (item) => !seenIds.has(item.id),
-        );
-        return [...current, ...appendedItems];
-      });
-      setNextCursor(response.nextCursor ?? null);
-    } catch (error) {
-      logger.error('Failed to load more global knowledge entries', error);
-      toast.error(
-        t('knowledge.loadMoreFailed', 'Failed to load more knowledge entries.'),
-      );
-    } finally {
-      setIsLoadingMore(false);
-    }
-  }, [
+  const {
+    query,
+    setQuery,
     assistantFilter,
-    debouncedQuery,
+    setAssistantFilter,
+    items,
+    assistants,
+    selectedId,
+    detail,
     isListLoading,
     isLoadingMore,
-    nextCursor,
-    t,
-  ]);
-
-  const handleRequestDelete = useCallback(() => {
-    if (!selectedItem || isDeleting) {
-      return;
-    }
-    setIsDeleteDialogOpen(true);
-  }, [isDeleting, selectedItem]);
-
-  const handleDelete = useCallback(async () => {
-    if (!selectedItem || isDeleting) {
-      return;
-    }
-
-    setIsDeleting(true);
-    try {
-      const response = await deleteGlobalKnowledge(selectedItem.id);
-      toast.success(t('knowledge.deleteSuccess', 'Knowledge entry deleted.'), {
-        description: t(
-          'knowledge.deleteSuccessDescription',
-          'Removed {{entities}} orphan entities and {{relationships}} orphan relationships.',
-          {
-            entities: response.orphanEntityCount,
-            relationships: response.orphanRelationshipCount,
-          },
-        ),
-      });
-      setDetail(null);
-      setSelectedId(null);
-      setIsDeleteDialogOpen(false);
-      setListRefreshToken((current) => current + 1);
-    } catch (error) {
-      logger.error('Failed to delete knowledge entry', {
-        id: selectedItem.id,
-        error,
-      });
-      toast.error(
-        t('knowledge.deleteFailed', 'Failed to delete knowledge entry.'),
-      );
-    } finally {
-      setIsDeleting(false);
-    }
-  }, [isDeleting, selectedItem, t]);
+    isDetailLoading,
+    isDeleting,
+    isDeleteDialogOpen,
+    setIsDeleteDialogOpen,
+    hasMoreItems,
+    isDetailOpen,
+    selectedItem,
+    entityNameById,
+    handleRefresh,
+    handleCloseDetail,
+    handleSelectItem,
+    handleLoadMore,
+    handleRequestDelete,
+    handleDelete,
+  } = useGlobalKnowledge();
 
   return (
     <div className="h-full bg-background p-6">
